refactor: Updated OptionMenu component to Radio
